Functional Burger Menu//CSS tweaks
Making the burger menu functional; appearing, disappearing, changing icon from a menu icon to a cross and back when menu is opened. Links to be added at a later date once other pages have been constructed.

Some small tweaks to css also to make type styling a little bit more modular, with the introduction of header and sub-header classes which can be re-used in differing contexts.