Refactored menu list display, improved history API usage
Managed to sort out some of the bugs with implementing the History API, and also cleaned up the way that options are displayed in the menu, added to the DOM, and given eventListeners.  Stored all pageDisplay functions in an object so they can be accessed more easily from within a loop.