Ajax working, no page redirection
Using the AJAX API for formSubmit I could make it so that when an email is sent using the contact form, the page doesn't reload or redirect. Hopefully this should give a cleaner user experience. In the future, I'd like to build my own backend with nodemailer but I want to get this project launched ASAP so this should work fin for now